backend/pyproject.toml
[project]
name = "kenbot"
version = "0.1.0"
description = "KenBot — automated Kenyan government portal navigator"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "django>=5.0,<6.0",
    "djangorestframework>=3.15",
    "channels>=4.1",
    "channels-redis>=4.2",
    "daphne>=4.1",
    # CORS
    "django-cors-headers>=4.3",
    # Agent frameworks
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "langgraph>=0.2",
    # Browser automation (Surveyor)
    "browser-use>=0.1",
    "playwright>=1.44",
    # LLM client
    "openai>=1.35",
    # Schema validation
    "pydantic>=2.7",
    # Task queue
    "celery>=5.4",
    "redis>=5.0",
    # Encryption
    "cryptography>=42.0",
    # JWT auth
    "djangorestframework-simplejwt>=5.3",
    # Utilities
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "django-stubs>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "N"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "kenbot.settings.development"
asyncio_mode = "auto"
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "kenbot.settings.development"
